fix :  캠페인 카테고리 별 결제 내역 조회 - 메서드 수정 및 환율 적용 시 소수점 반올림 적용

+++ src/main/java/com/keypoint/keypointtravel/campaign/dto/request/createRequest/BudgetInfo.java
@@ -19,7 +19,7 @@ public class BudgetInfo {
 
     @NotEmpty(message = "금액을 입력해 주세요")
     @Min(value = 1, message = "금액은 1원 이상이어야 합니다.")
-    private Long amount;
+    private float amount;
 
     private CurrencyType currency;
 }
